refactor(composer): extract slash_popup + history_nav as composer/ submodules (closes #1187)
Closes #1187. Pure refactor, zero user-visible change. Delivers the
composer-locality goal from the original textarea epic (#1178 PR 3
plan).

## What moved where

| From | To | What |
|---|---|---|
| `completer.rs` | `composer/slash_popup.rs` | `SLASH_COMMANDS` const |
| `completer.rs` (`complete_slash` body) | `composer/slash_popup.rs::matches_for` | Slash matching algorithm (no cycling state) |
| `widgets/slash_menu.rs` (whole file) | `composer/slash_popup.rs` | `SlashCommand`, `from_input`, `build_menu_lines`, all 4 tests |
| `tui_context/events.rs` | `composer/history_nav.rs` | `history_up_index`, `history_down_index`, all 7 tests |

Net diff:
- **-235 LoC** at old call sites (completer.rs -47, events.rs -82,
  widgets/slash_menu.rs -94 deleted, widgets/mod.rs -1, others -11)
- **+~370 LoC** in two new focused modules (~250 slash_popup +
  ~120 history_nav, of which ~150 LoC is doc comments + tests)

## Why

Following the rationale in #1187:

- **Locality** \u2014 composer-related state (textarea, slash, history,
  paste_burst) all live under one module tree, easier to reason about.
- **Future syncs** \u2014 when codex evolves their slash/history logic,
  our diff is easier to review when our shape mirrors theirs.
- **Test isolation** \u2014 history nav state + slash matching logic are
  easier to unit-test as standalone modules than embedded in a
  700-line event handler.

## API stability

`completer.rs` re-exports `SLASH_COMMANDS` from
`composer::slash_popup` so existing callers (and any external doc-
tests) keep working without import changes:

    pub use crate::composer::slash_popup::SLASH_COMMANDS;

The `InputCompleter` cycling state machine stays put \u2014 just its
matching helper is now a free function in slash_popup. This keeps the
`completer.complete()` Tab-completion API surface unchanged for the
two remaining callers (`tui_handlers_inference`, `tui_context`).

## Bonus: 4 new unit tests

The new `matches_for` free function added 4 tests against the
production `SLASH_COMMANDS` table:

- unique-prefix returns one
- complete-command returns nothing (no self-completion)
- no-match returns empty
- ambiguous prefix returns all sharing the prefix

These double as a smoke test that the production slash table still
contains the commands users expect.

## Quality gates

- `cargo fmt --all --check`: clean
- `cargo clippy -p koda-cli --all-targets --features koda-core/test-support -- -D warnings`: clean
- `cargo test -p koda-cli --lib`: 576/576 passing (was 572 baseline; +4 from new tests)
- `cargo test --workspace`: **2353/2353 passing** (issue spec asked for 2349+)
- Zero user-visible behavior change

## Refs

- Parent epic: #1116
- Textarea epic: #1178 (closed)
- Sibling: #1186 (paste_burst wiring) \u2014 PR #1189, lands independently
